refactor(Line): 라인의 생성자 변경, 검증로직 추가
