+++ src/main/java/dto/LineInfo.java
@@ -5,7 +5,7 @@ import model.Line;
 
 public record LineInfo(List<Boolean> lineInfo) {
     public static LineInfo from(Line line) {
-        List<Boolean> lineInfo = line.getLineInfo();
+        List<Boolean> lineInfo = line.getExistFlags();
         return new LineInfo(lineInfo);
     }
 }
