+++ src/main/java/formatter/LineFormatter.java
@@ -5,15 +5,17 @@ import java.util.List;
 public class LineFormatter {
 
     public static String format(List<Boolean> line) {
-        // TODO : 끝내주는 리팩토링 하기
         String formattedLine = " ".repeat(4) + "|";
         for (Boolean isPath : line) {
-            if (isPath) {
-                formattedLine += "-".repeat(5) + "|";
-                continue;
-            }
-            formattedLine += " ".repeat(5) + "|";
+            formattedLine += formatPath(isPath);
         }
         return formattedLine;
+    }
+
+    private static String formatPath(Boolean isPath) {
+        if (isPath) {
+            return "-".repeat(5) + "|";
+        }
+        return " ".repeat(5) + "|";
     }
 }
