+++ src/main/java/model/path/Path.java
@@ -1,4 +1,4 @@
-package model;
+package model.path;
 
 public enum Path {
     EXIST(true),
